requirements.txt
streamlit
pandas
numpy
scikit-learn
tensorflow
pillow
opencv-python
transformers
torch
torchvision
nltk
requests
plotly
bcrypt
python-dotenv
huggingface-hub
datasets
sentencepiece
ultralytics
hf_xet